Create directory structure and model files
