+++ mcp_server/src/graphiti_mcp_server.py
@@ -36,6 +36,20 @@ from pydantic import BaseModel, Field
 
 from src.config_loader import config_loader
 from src.ollama_client import OllamaClient
+# Import model definitions from the models package
+from src.models import (
+    Requirement,
+    Preference,
+    Procedure,
+    ENTITY_TYPES,
+    ErrorResponse,
+    SuccessResponse,
+    NodeResult,
+    NodeSearchResponse,
+    FactSearchResponse,
+    EpisodeSearchResponse,
+    StatusResponse,
+)
 
 load_dotenv()
 
@@ -48,125 +62,6 @@ DEFAULT_EMBEDDER_MODEL = 'nomic-embed-text'
 # Decrease this if you're experiencing 429 rate limit errors from your LLM provider.
 # Increase if you have high rate limits.
 SEMAPHORE_LIMIT = int(os.getenv('SEMAPHORE_LIMIT', 10))
-
-
-class Requirement(BaseModel):
-    """A Requirement represents a specific need, feature, or functionality that a product or service must fulfill.
-
-    Always ensure an edge is created between the requirement and the project it belongs to, and clearly indicate on the
-    edge that the requirement is a requirement.
-
-    Instructions for identifying and extracting requirements:
-    1. Look for explicit statements of needs or necessities ("We need X", "X is required", "X must have Y")
-    2. Identify functional specifications that describe what the system should do
-    3. Pay attention to non-functional requirements like performance, security, or usability criteria
-    4. Extract constraints or limitations that must be adhered to
-    5. Focus on clear, specific, and measurable requirements rather than vague wishes
-    6. Capture the priority or importance if mentioned ("critical", "high priority", etc.)
-    7. Include any dependencies between requirements when explicitly stated
-    8. Preserve the original intent and scope of the requirement
-    9. Categorize requirements appropriately based on their domain or function
-    """
-
-    project_name: str = Field(
-        ...,
-        description='The name of the project to which the requirement belongs.',
-    )
-    description: str = Field(
-        ...,
-        description='Description of the requirement. Only use information mentioned in the context to write this description.',
-    )
-
-
-class Preference(BaseModel):
-    """A Preference represents a user's expressed like, dislike, or preference for something.
-
-    Instructions for identifying and extracting preferences:
-    1. Look for explicit statements of preference such as "I like/love/enjoy/prefer X" or "I don't like/hate/dislike X"
-    2. Pay attention to comparative statements ("I prefer X over Y")
-    3. Consider the emotional tone when users mention certain topics
-    4. Extract only preferences that are clearly expressed, not assumptions
-    5. Categorize the preference appropriately based on its domain (food, music, brands, etc.)
-    6. Include relevant qualifiers (e.g., "likes spicy food" rather than just "likes food")
-    7. Only extract preferences directly stated by the user, not preferences of others they mention
-    8. Provide a concise but specific description that captures the nature of the preference
-    """
-
-    category: str = Field(
-        ...,
-        description="The category of the preference. (e.g., 'Brands', 'Food', 'Music')",
-    )
-    description: str = Field(
-        ...,
-        description='Brief description of the preference. Only use information mentioned in the context to write this description.',
-    )
-
-
-class Procedure(BaseModel):
-    """A Procedure informing the agent what actions to take or how to perform in certain scenarios. Procedures are typically composed of several steps.
-
-    Instructions for identifying and extracting procedures:
-    1. Look for sequential instructions or steps ("First do X, then do Y")
-    2. Identify explicit directives or commands ("Always do X when Y happens")
-    3. Pay attention to conditional statements ("If X occurs, then do Y")
-    4. Extract procedures that have clear beginning and end points
-    5. Focus on actionable instructions rather than general information
-    6. Preserve the original sequence and dependencies between steps
-    7. Include any specified conditions or triggers for the procedure
-    8. Capture any stated purpose or goal of the procedure
-    9. Summarize complex procedures while maintaining critical details
-    """
-
-    description: str = Field(
-        ...,
-        description='Brief description of the procedure. Only use information mentioned in the context to write this description.',
-    )
-
-
-ENTITY_TYPES: dict[str, BaseModel] = {
-    'Requirement': Requirement,  # type: ignore
-    'Preference': Preference,  # type: ignore
-    'Procedure': Procedure,  # type: ignore
-}
-
-
-# Type definitions for API responses (using Pydantic for MCPO compatibility)
-class ErrorResponse(BaseModel):
-    error: str
-
-
-class SuccessResponse(BaseModel):
-    message: str
-
-
-class NodeResult(BaseModel):
-    uuid: str
-    name: str
-    summary: str
-    labels: List[str]
-    group_id: str
-    created_at: str
-    attributes: dict[str, Any]
-
-
-class NodeSearchResponse(BaseModel):
-    message: str
-    nodes: List[NodeResult]
-
-
-class FactSearchResponse(BaseModel):
-    message: str
-    facts: List[dict[str, Any]]
-
-
-class EpisodeSearchResponse(BaseModel):
-    message: str
-    episodes: List[dict[str, Any]]
-
-
-class StatusResponse(BaseModel):
-    status: str
-    message: str
 
 
 def create_azure_credential_token_provider() -> Callable[[], str]:
